took out community and made separate pages
