+++ bifrost/app/case-studies/page.tsx
@@ -1,4 +1,4 @@
-import { Customers } from "../community/customers";
+import { Customers } from "../customers/page";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
