+++ bifrost/app/integrations/layout.tsx
@@ -1,0 +1,9 @@
+import { Layout } from "@/app/components/Layout";
+
+export default function RootLayount({
+    children,
+}: {
+    children: React.ReactNode;
+}) {
+    return <Layout>{children}</Layout>;
+}
