+++ bifrost/components/layout/footer.tsx
@@ -4,7 +4,7 @@ import { SVGProps } from "react";
 import Image from "next/image";
 import { usePathname } from "next/navigation";
 
-interface FooterProps {}
+interface FooterProps { }
 
 const meta = {
   social: [
@@ -53,11 +53,10 @@ const Footer = () => {
 
   return (
     <footer
-      className={`grid grid-cols-2 md:grid-cols-5 gap-y-8 pl-8 md:pl-0 md:justify-items-center items-start ${
-        path === "/"
-          ? " bg-white text-slate-700 fill-slate-700 stroke-slate-700 py-16"
-          : "bg-inherit text-black/60 fill-[#5D6673] stroke-[#5D6673] py-6"
-      }`}
+      className={`grid grid-cols-2 md:grid-cols-5 gap-y-8 pl-8 md:pl-0 md:justify-items-center items-start ${path === "/"
+        ? " bg-white text-slate-700 fill-slate-700 stroke-slate-700 py-16"
+        : "bg-inherit text-black/60 fill-[#5D6673] stroke-[#5D6673] py-6"
+        }`}
     >
       <div className="flex flex-col items-start font-light text-sm tracking-wide gap-1 col-span-2 md:col-span-1 justify-self-start md:justify-self-center">
         <Image
@@ -188,9 +187,6 @@ const Footer = () => {
             target="_blank"
           >
             Stats
-          </a>
-          <a className="hover:underline" href="/community" target="_blank">
-            Community
           </a>
           <a className="hover:underline" href="/changelog" target="_blank">
             Changelog
